rename files manga.dart and chapter.dart
