+++ lib/modules/manga/reader/mixins/chapter_controller_mixin.dart
@@ -4,7 +4,7 @@ import 'package:mangayomi/models/chapter.dart';
 import 'package:mangayomi/models/history.dart';
 import 'package:mangayomi/models/manga.dart';
 import 'package:mangayomi/models/settings.dart';
-import 'package:mangayomi/utils/extensions/manga.dart';
+import 'package:mangayomi/utils/extensions/manga_extensions.dart';
 
 /// Shared navigation and history logic used by [ReaderController],
 /// [NovelReaderController], and [AnimeStreamController].
