+++ lib/modules/updates/widgets/update_chapter_list_tile_widget.dart
@@ -7,7 +7,7 @@ import 'package:mangayomi/models/chapter.dart';
 import 'package:mangayomi/modules/widgets/custom_extended_image_provider.dart';
 import 'package:mangayomi/utils/constant.dart';
 import 'package:mangayomi/modules/manga/download/download_page_widget.dart';
-import 'package:mangayomi/utils/extensions/chapter.dart';
+import 'package:mangayomi/utils/extensions/chapter_extensions.dart';
 import 'package:mangayomi/utils/headers.dart';
 
 class UpdateChapterListTileWidget extends ConsumerWidget {
