+++ lib/utils/extensions/manga_extensions.dart
